Share messages related data types between addin and backend
