+++ AppBundles/DrawingsListPlugin/Properties/AssemblyInfo.cs
@@ -1,10 +1,10 @@
 ﻿using System.Reflection;
 using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("DrawingsListPlugin")]
+[assembly: AssemblyTitle("Adoption Data Checker adding")]
 [assembly: AssemblyDescription("")]
 
 [assembly: Guid("f1c9d7e2-53a8-4e4e-af9e-931ca891715d")]
 
-[assembly: AssemblyVersion("1.0.0.6")]
-[assembly: AssemblyFileVersion("1.0.0.6")]
+[assembly: AssemblyVersion("1.0.0.8")]
+[assembly: AssemblyFileVersion("1.0.0.8")]
