+++ Shared/Message.cs
@@ -1,0 +1,48 @@
+﻿/////////////////////////////////////////////////////////////////////
+// Copyright (c) Autodesk, Inc. All rights reserved
+// Written by Forge Design Automation team for Inventor
+//
+// Permission to use, copy, modify, and distribute this software in
+// object code form for any purpose and without fee is hereby granted,
+// provided that the above copyright notice appears in all copies and
+// that both that copyright notice and the limited warranty and
+// restricted rights notice below appear in all supporting
+// documentation.
+//
+// AUTODESK PROVIDES THIS PROGRAM "AS IS" AND WITH ALL FAULTS.
+// AUTODESK SPECIFICALLY DISCLAIMS ANY IMPLIED WARRANTY OF
+// MERCHANTABILITY OR FITNESS FOR A PARTICULAR USE.  AUTODESK, INC.
+// DOES NOT WARRANT THAT THE OPERATION OF THE PROGRAM WILL BE
+// UNINTERRUPTED OR ERROR FREE.
+/////////////////////////////////////////////////////////////////////
+
+// Data types from this file are shared between .NET 4.7+ and netcore projects,
+// so we need to have different attributes for Newtonsoft and netcore Json libraries.
+
+#if NETCOREAPP
+using JsonProperty = System.Text.Json.Serialization.JsonPropertyNameAttribute;
+#else
+using JsonProperty = Newtonsoft.Json.JsonPropertyAttribute;
+#endif
+
+namespace Shared
+{
+    /// <summary>
+    /// Message severity.
+    /// </summary>
+    public enum Severity
+    {
+        Info = 0,
+        Warning = 1,
+        Error = 2
+    }
+
+    public class Message
+    {
+        [JsonProperty("text")]
+        public string Text { get; set; }
+
+        [JsonProperty("severity")]
+        public Severity Severity { get; set; }
+    }
+}
